Move ui theme files to presentation layer
