+++ app/src/main/java/org/groundplatform/android/PermissionDeniedDialog.kt
@@ -33,8 +33,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import org.groundplatform.android.presentation.ui.theme.AppTheme
 import org.groundplatform.android.ui.compose.HyperlinkText
-import org.groundplatform.android.ui.theme.AppTheme
 import timber.log.Timber
 
 @Composable
