+++ app/src/main/java/org/groundplatform/android/presentation/ui/theme/Color.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.groundplatform.android.ui.theme
+package org.groundplatform.android.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
