+++ app/src/main/java/org/groundplatform/android/ui/datacollection/components/InstructionsDialog.kt
@@ -33,7 +33,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import org.groundplatform.android.ExcludeFromJacocoGeneratedReport
 import org.groundplatform.android.R
-import org.groundplatform.android.ui.theme.AppTheme
+import org.groundplatform.android.presentation.ui.theme.AppTheme
 
 @Composable
 fun InstructionsDialog(iconId: Int, stringId: Int) {
