+++ app/src/main/java/org/groundplatform/android/ui/home/UserDetailsDialog.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import org.groundplatform.android.ExcludeFromJacocoGeneratedReport
 import org.groundplatform.android.R
 import org.groundplatform.android.model.User
-import org.groundplatform.android.ui.theme.AppTheme
+import org.groundplatform.android.presentation.ui.theme.AppTheme
 
 @Composable
 fun UserDetailsDialog(user: User, signOutCallback: () -> Unit, dismissCallback: () -> Unit) {
