+++ app/src/main/java/org/groundplatform/android/util/ComposeExt.kt
@@ -21,7 +21,7 @@ import android.view.ViewGroup
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
-import org.groundplatform.android.ui.theme.AppTheme
+import org.groundplatform.android.presentation.ui.theme.AppTheme
 
 /**
  * Renders a composable function within the root view of a Fragment.
